refactor(de-monolith): ADR-0035 Stage 0a — extract kirra-policy-types leaf crate
First slice of the layered decomposition (ADR-0035). Extract the
command-classification policy types — OperationalCommand + classify_http_command
+ the private path allowlists (is_write_state_path / is_system_mutation_path) —
VERBATIM out of the root crate's src/gateway/policy.rs into a new
zero-dependency leaf crate `kirra-policy-types`.

Behaviour-preserving pure move:
- src/gateway/policy.rs becomes a `pub use kirra_policy_types::{OperationalCommand,
  classify_http_command}` re-export shim, so every existing
  crate::gateway::policy::* path (posture gate, industrial adapters, ~10
  consumers) resolves unchanged.
- The pure classifier tests (12) move with the code into the new crate; the ONE
  cross-layer test (test_unrecognized_write_path_denied_in_all_postures, which
  exercises should_route_command from posture_cache — a layer ABOVE this leaf in
  the DAG) stays in the root shim.

This is the gating prerequisite for Stage 1 (kirra-industrial): the industrial
adapters import gateway::policy::OperationalCommand, and a new leaf crate can't
depend back on the root — so the classification types had to move down first.

Mechanics proven for the whole program: new workspace member + root path-dep,
re-export shim, quality-guardrails (policy.rs 447→53, satisfied), orphan-core
gate green (the shim keeps real `use` consumers; the crate has no pub mod to
orphan), fmt, and BOTH lockfiles updated minimally (main +5, fuzz +5 — only the
new crate added, no unrelated churn). 12 policy-types tests + 16 root policy
tests (incl. the cross-layer shim test) green.
